requirements.txt
pysdl2
pysdl2-dll
pygame
click
python-util
keyboard